Rename to whole_body_state_rviz_plugin
