+++ README.md
@@ -3,7 +3,7 @@ State Rviz plugins
 
 ## <img align="center" height="20" src="https://i.imgur.com/vAYeCzC.png"/> Introduction
 
-This repository includes Rviz plugins to display the whole-body state and trajectory of a robot. These plugins use dedicated ROS messages defined in [state_msgs](https://github.com/cmastalli/state_msgs) package. To process efficiently the messages, e.g. robot display without TF broadcasting, the plugins use [Pinocchio](https://github.com/stack-of-tasks/pinocchio) library.
+This repository includes Rviz plugins to display the whole-body state and trajectory of a robot. These plugins use dedicated ROS messages defined in [whole_body_state_msgs](https://github.com/cmastalli/whole_body_state_msgs) package. To process efficiently the messages, e.g. robot display without TF broadcasting, the plugins use [Pinocchio](https://github.com/stack-of-tasks/pinocchio) library.
 
 ![output](https://user-images.githubusercontent.com/3601935/89519116-c6b45600-d7d3-11ea-89a0-fc8df97df2f0.gif)
 
@@ -28,7 +28,7 @@ All visuals are configurable through Rviz GUI. For example, it is possible to co
 
 1. Installation pinocchio from any source (ros / robotpkg binaries or source)
 
-2. Building the state_msgs and these packages in your catkin workspace:
+2. Building the whole_body_state_msgs and these packages in your catkin workspace:
     ```bash
 	cd your_ros_ws/
 	catkin build #catkin_make
