+++ include/whole_body_state_rviz_plugin/PointVisual.h
@@ -6,11 +6,8 @@
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
 
-#ifndef STATE_RVIZ_PLUGIN_LINE_VISUAL_H
-#define STATE_RVIZ_PLUGIN_LINE_VISUAL_H
-
-#include <Eigen/Dense>
-#include <rviz/properties/quaternion_property.h>
+#ifndef WHOLE_BODY_STATE_RVIZ_PLUGIN_POINT_VISUAL_H
+#define WHOLE_BODY_STATE_RVIZ_PLUGIN_POINT_VISUAL_H
 
 namespace Ogre {
 class Vector3;
@@ -18,37 +15,34 @@ class Quaternion;
 } // namespace Ogre
 
 namespace rviz {
-class Arrow;
+class Shape;
 }
 
-namespace state_rviz_plugin {
+namespace whole_body_state_rviz_plugin {
 
 /**
- * @class LineVisual
- * @brief Visualizes 3d line
- * Each instance of LineVisual represents the visualization of
- * a single arrow data. Currently it just shows an line with
- * the initial and final points
+ * @class PointVisual
+ * @brief Visualizes 3d point
+ * Each instance of PointVisual represents the visualization of a single
+ * Ogre::Vector3 data. Currently it just shows a sphere in the point position
  */
-class LineVisual {
+class PointVisual {
 public:
   /**
    * @brief Constructor that creates the visual stuff and puts it into the scene
    * @param scene_manager  Manager the organization and rendering of the scene
-   * @param parent_node  Represent the arrow as node in the scene
+   * @param parent_node    Represent the point as node in the scene
    */
-  LineVisual(Ogre::SceneManager *scene_manager, Ogre::SceneNode *parent_node);
+  PointVisual(Ogre::SceneManager *scene_manager, Ogre::SceneNode *parent_node);
 
   /** @brief Destructor that removes the visual stuff from the scene */
-  ~LineVisual();
+  ~PointVisual();
 
   /**
-   * @brief Configure the visual to show the line from their points
-   * @param initial_point  Initial point of the line
-   * @param final_point    Final point of the line
+   * @brief Configure the visual to show the point
+   * @param point  Point position
    */
-  void setArrow(const Ogre::Vector3 &initial_point,
-                const Ogre::Vector3 &final_point);
+  void setPoint(const Ogre::Vector3 &point);
 
   /**
    * @brief Set the position of the coordinate frame
@@ -72,17 +66,14 @@ public:
   void setColor(float r, float g, float b, float a);
 
   /**
-   * @brief Set the parameters for this arrow
-   * @param shaft_diameter  Diameter of the arrow's shaft
-   * @param head_length     Length of the arrow's head
-   * @param head_diameter   Diameter of the arrow's head
+   * @brief Set the radius of the point
+   * @param r  Radius value
    */
-  void setProperties(float shaft_diameter, float head_length = 0.,
-                     float head_diameter = 0.);
+  void setRadius(float r);
 
 private:
-  /** @brief The object implementing the arrow */
-  rviz::Arrow *arrow_;
+  /** @brief The object implementing the point circle */
+  rviz::Shape *point_;
 
   /** @brief A SceneNode whose pose is set to match the coordinate frame */
   Ogre::SceneNode *frame_node_;
@@ -92,10 +83,10 @@ private:
    */
   Ogre::SceneManager *scene_manager_;
 
-  /** @brief Distance of the line */
-  float distance_;
+  /** @brief Radius value */
+  float radius_;
 };
 
-} // namespace state_rviz_plugin
+} // namespace whole_body_state_rviz_plugin
 
-#endif // STATE_RVIZ_PLUGIN_LINE_VISUAL_H
+#endif // WHOLE_BODY_STATE_RVIZ_PLUGIN_POINT_VISUAL_H
